Update base image to a supported version (22.04)
It was a lovely time back in 2018 where pandemic was not in the
vocabulary of the common man. Many people continued to be stuffed
into tubes on their way to work, blissfully unaware of the impending
chaos of 2020. So lets skip pass those years fully into 2022 where
nothing else of note happened...